Add simple CSV file uploading functionality
This application will be dealing with lots of CSV file to import
the object states, simplest way could be we instantly upload the
file and then import those data in the fly, and most likely this
will work great for the smaller dataset.

But in the long run this might not be the case, we might need to
deal with large CSV file with huge number of data to import & we
might also want to refer back the files that was used to import
any specific entry and that's where the above approach might not
work best.

So, a better approach could be to upload the CSV file and store
it in the application and then import it from there. This way we
can refer back to the file and in the future we might also think
about processing those in the background