Import all valid CSV data as Tg Object
In the previous commit we've added the functionality to upload
a CSV file and the next step is to process and store those data
properly. But before importing those data here are the some of
the major considerations.

The happiest path, user uploads a CSV with all of the valid data
and in that case it is safe to import those data and update the
CSV file status to `imported`.

Second consideration, user might upload a CSV file with partial
valid data with some invalidate or duplicate entry, and in that
case we might import the valid entry but also keep track of the
invalid entry so later we might also offer the user to download
those data and fix those validation error and re-upload.

And this commit consider those above scenario & add the changes
to import user uploaded data properly with the tracking option
for those invalid entry.

Future Improvements:

In this current version, we are reading a file from a remote io
and it should work fine for our purpose for now, but if it does
any harm on performance then we might consider downloading the
file first and then do the CSV parsing.