Add Object State tracking functionality
In the previous commits, we have allowed the user to upload and
import CSV data to our application and the next step is to allow
them to track the state for any specific object. Here are some
of the consideration for this feature:

User tries to track an object with all of the valid information
and they are expected to see the object state changes during that
time frame and it should reflect the recent timestamp changes as
latest state for an attribute.

Secondly, user tries to track an object that did not existed in
that time. In this scenario they should see non-existence objec
state related messages.

And the last, user tries to track an object without any valid
tracking related information, and in this scenario they should
see that they are missing to track any object.

This commit took all of the above considerations and implement
the proper functionality to track an object state properly.

Helpful Resources:

* http://blog.faraday.io/how-to-aggregate-jsonb-in-postgres/
* https://postgresql.org/docs/9.6/static/sql-createaggregate.html