Optimizing CSV data importing procedure
Currently, we are trying to process the CSV data instantly, which
is not ideal for a large sets of data, and we are also not doing
any kind of validation for the json data, which might introduce
some problem in the long run.

This commit takes those into consideration and create a process
to send the data parsing work in the background, and it also add
some parsing level validation to ensure we only store valid data

To keep the scope small we're not doing any validation in regards
to the csv file and data structure in it, so that's something we
can do as a future improvement to this application.