Optimize JSON aggregation related SQL Query
In the `TgObject` model we have an interface that does the job to
aggregate the object changes and finally we show that to the user

In the underneath we're running sub-query to sort those changes
and then pass it to the aggregation function. This is probably
fine and might not effect anything at all.

But looking at the query plan it seems like there is a better a
way to do it and that's what we've implemented in this query.