Add support for non-standard csv format
Base on the RFC standard double quotes should be escaped by a
preceding double quotes, but based on the provided sample looks
like we have some no-standard format to represent the changes of
an object as json string.

Ideally, it might be a good idea to avoid this, as most of the
common tools won't have direct support for this type of format,
but if that's what we need to support then this PR might come
in handy.

Couple of consideration here though, before parsing we might
replaces those `\"` with the proper format, but that might be
a performance issue for bigger file.

Another approach could be change the `quote_char` something
else so the csv parser does not throw any error instantly,
and then manually unscape those object changes, but this will
add some limitations.

So, considering both of the scenario, the first one seems
more suitable, so if this what we want then please feel
free to merge this PR.

Reference: https://tools.ietf.org/html/rfc4180#section-2