Add correct status for invalid CSV file
Currently, we are assuming to provide the csv file with correct
format, but sometime that might not be the case and our importer
throws an `CSV::MalformedCSVError` error for those scenario but
the problem is that we do not clearly specify the correct status
for that upload.

This commit improves it a bit, and for any reason if our importer
detect an `CSV::MalformedCSVError` error then it will mark the
CSV file as invalid format.